Measure the weight entirely for automatic
So far, the target bean amount was approximated by grinding for 6 seconds, and then approaching it with 150 ms steps.
Apart from being slower and loud, smaller bean amounts were not supported.

Now the mill will try to reach the desired bean amount as fast as possible, and if needed, grind for another 150 ms.